Refactor authentication provider imports and fix import paths in index.tsx
